migration to rust8 Pre

+++ test/libcss/dump.rs
@@ -2239,7 +2239,7 @@ fn dump_bytecode(style:@mut css_style, lwc_ref:&mut ~lwc, ptr:&mut ~str, depth:u
 fn dump_number(val: i32 , ptr: &mut ~str){
     debug!("Entering: dump_number");
     if css_int_to_fixed((val >> 10) as int) == val {
-        ptr.push_str( fmt!("%i" , val as int >> 10 as int));
+        ptr.push_str( fmt!("%i" , val as int >> 10));
     }
     else {
         dump_css_fixed(val , ptr);
